fix float sur chmaps coordonne

+++ database/migrations/2025_05_01_180033_create_expeditions_table.php
@@ -13,8 +13,8 @@ return new class extends Migration {
         Schema::create('expeditions', function (Blueprint $table) {
             $table->id();
             $table->string('adresse');
-            $table->integer('longitude');
-            $table->integer('latitude');
+            $table->float('longitude');
+            $table->float('latitude');
             $table->string('tel_expedition')->nullable();
             $table->timestamps();
         });
